Added player, camera, gamemanager
Player and camera script is added, gamemanger is added but there's an issue with Pause/Unpause not passing the menu to the tempvariable. Need to leave for an appointment will get back to this later unless someone else has time to debug. Player movement needs to be tuned but is working and camera seems to be working, didn't test whether the player is moving with the camera (tilt) or not and the player is "popping up" above the floor on play.