Settings Menu
Still working on the settings, I did get the animation/transition to work for the pause menu, but got hung up on the settings not turning on, now there seems to be another issue with transparency even though the alpha value is 1. I will continue tomorrow but I think once I get over this issue and get the navigation from screen to screen the rest will be easy.