Finally fixed and back to almost where I was before
finally fixed the menus (sort of) the pause and settings able to be called and put away, the placeholders for the HUD are back up and the other menu screens are in place, they just need to be hooked up. lost some icons, but those should be able to be put back in.