pause fix, health bar ease and crit fix
pause changed back to timescale for the sake of time and the fact that DOTween doesn't work ok webGL so I will undo the animations or put in a check for webGL. Health bar ease wasn't working because the speeds were switched, and crit update needed to be changed to game object. weapon pickups aren't working but health is. No feedback on health pickup yet